refactor : Player, Dealer 간 중복 코드를 조합 방식으로 제거
