+++ src/main/java/model/participant/Player.java
@@ -1,25 +1,51 @@
 package model.participant;
 
 import java.util.List;
+import java.util.stream.IntStream;
 import model.deck.Card;
+import model.deck.Deck;
 
-public final class Player extends Participant {
+public final class Player {
+    private static final int INITIAL_DEAL_CARD_COUNT = 2;
+
     private final String name;
+    private final ParticipantHand participantHand;
 
     public Player(final String name) {
-        super();
         this.name = name;
+        this.participantHand = new ParticipantHand();
     }
 
-    @Override
+    public void receiveCard(final Card card) {
+        participantHand.add(card);
+    }
+
+    public boolean isBurst() {
+        return participantHand.checkBurst();
+    }
+
+    public void dealInitialCards(final Deck deck) {
+        IntStream.range(0, INITIAL_DEAL_CARD_COUNT).forEach(
+                i -> receiveCard(deck.pick())
+        );
+    }
+
+    public int calculateFinalScore() {
+        return participantHand.calculateFinalScore();
+    }
+
     public List<Card> openInitialDeal() {
         return participantHand.getCards();
     }
 
-    @Override
     public boolean canHit() {
         return !isBurst();
     }
+
+    public List<Card> getHandCards() {
+        return participantHand.getCards();
+    }
+
 
     public String getName() {
         return name;
